Added favicon to tv application; broken down advisor student queue and information into 2 components; added comments for confluence purposes; removed back button from degree level screen in kiosk;
