+++ frontend/tv_web_application/AdvisorStudentQueue.js
@@ -1,9 +1,7 @@
 import React from "react";
-import { View, StyleSheet, Image, Dimensions, Text } from "react-native";
+import { View, StyleSheet, Image, Dimensions } from "react-native";
 // import axios from "axios";
 import AdvisorStudentCard from "./AdvisorStudentCard";
-import { images, util } from "./Utility";
-import { Video, AVPlaybackStatus } from "expo-av";
 
 let advisorsWithStudents = [
   {
@@ -51,25 +49,15 @@ let advisorsWithStudents = [
   },
 ];
 
-let mediaContent = {
-  title:
-    "Golisano College School of Information Presents: Jump Starting Your Career",
-  firstLine: "November 09, 2021",
-  secondLine: "3:30 pm - 4:30 pm",
-  thirdLine: "Virtual",
-  media: "http://d23dyxeqlo5psv.cloudfront.net/big_buck_bunny.mp4",
-  //   media: "https://source.unsplash.com/830x350?technology",
-  mediaType: "video",
-  //   mediaType: "image",
-};
-
+/**
+ * This class is responsible for showing the advisors and the students in thier queue
+ * Author: Raghul Krishnan
+ */
 export default class AdvisorStudentQueue extends React.Component {
-  videoRef;
   constructor(props) {
     super(props);
     this.state = {
       advisorStudentList: advisorsWithStudents,
-      mediaContent: mediaContent,
     };
     this.getData();
   }
@@ -120,56 +108,6 @@ export default class AdvisorStudentQueue extends React.Component {
           source={require("./assets/bg_right_triangle.png")}
           style={styles.rightTraingle}
         />
-        <View style={styles.newsContainer}>
-          <View style={styles.newsContentContainer}>
-            <View style={styles.floatingHeader}>
-              <Text style={styles.newsHeaderText}>What's Going On?</Text>
-            </View>
-            <Text style={styles.newsTitleText}>
-              {this.state.mediaContent.title}
-            </Text>
-            <View style={styles.informationContainer}>
-              {Object.keys(this.state.mediaContent).map((key) => {
-                return (
-                  !key.startsWith("media") &&
-                  key != "title" && (
-                    <Text style={styles.informationText} key={key}>
-                      {this.state.mediaContent[key]}
-                    </Text>
-                  )
-                );
-              })}
-            </View>
-          </View>
-          {this.state.mediaContent.mediaType == "image" && (
-            <Image
-              source={this.state.mediaContent.media}
-              style={{
-                width: Dimensions.get("window").width * 0.5 - 165,
-              }}
-            />
-          )}
-          {this.state.mediaContent.mediaType == "video" && (
-            <Video
-              style={{
-                width: Dimensions.get("window").width * 0.5 - 165,
-              }}
-              source={{
-                uri: this.state.mediaContent.media,
-              }}
-              useNativeControls
-              resizeMode="cover"
-              isLooping
-              shouldPlay={true}
-            />
-          )}
-          <Image
-            source={images.rit_gccis_image}
-            style={{
-              width: 289,
-            }}
-          />
-        </View>
       </View>
     );
   }
@@ -178,7 +116,6 @@ export default class AdvisorStudentQueue extends React.Component {
 const styles = StyleSheet.create({
   topView: {
     flex: 1,
-    flexDirection: "column",
   },
   leftTraingle: {
     position: "absolute",
@@ -200,52 +137,5 @@ const styles = StyleSheet.create({
     width: 310,
     height: 350,
     zIndex: -1,
-  },
-  floatingHeader: {
-    flexDirection: "row",
-    zIndex: 10,
-  },
-  newsContainer: {
-    flex: 1,
-    bottom: 0,
-    position: "absolute",
-    left: 0,
-    borderWidth: 6,
-    borderColor: util.rit_primary_color,
-    flexDirection: "row",
-    maxHeight: Dimensions.get("window").height * 0.32,
-  },
-  newsContentContainer: {
-    position: "relative",
-    bottom: 40,
-    left: 20,
-    marginStart: 24,
-    flexDirection: "column",
-    marginRight: 32,
-    flexWrap: "wrap",
-    width: Dimensions.get("window").width * 0.4,
-  },
-  newsHeaderText: {
-    padding: 12,
-    backgroundColor: util.rit_primary_color,
-    borderRadius: 4,
-    borderBottomLeftRadius: 16,
-    color: "white",
-    fontSize: 32,
-    fontWeight: "bold",
-  },
-  newsTitleText: {
-    marginTop: 24,
-    fontSize: 24,
-    fontWeight: "bold",
-    flexWrap: "wrap",
-  },
-  informationContainer: {
-    marginTop: 56,
-    fontWeight: "bold",
-    fontSize: 24,
-  },
-  informationText: {
-    fontSize: 24,
   },
 });
